// src/app/models/model-auditoria.ts
export class ModelAuditoria {
    USUREG?: string;
    NOMAPE?: string;
    PAGINA?: string;
    UA?: string;
    DISPOSITIVO?: string;
    MOBILE?: string;
    PHONE?: string;
    TABLET?: string;
    OS?: string;
    USERAGENT?: string;
    IP_PUBLICA?: string;
    LATITUDE?: string;
    LONGITUDE?: string;
    APLICACION?: string;
    TIPOAPLICACION?: string;
    CODOPCION?: string;
}